requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.1